Combine the dialog and autosave dialog services. Opening an autosave dialog should be as simple as calling a different function to open a dialog with a form. Also, the controller no longer serves any purpose since the dialog component now handles the form.
