+++ source/services/dialog/dialogTypes.ts
@@ -1,0 +1,59 @@
+'use strict';
+
+import * as angular from 'angular';
+
+export interface IDialogCloseHandler {
+	(explicit: boolean): boolean;
+}
+
+export interface IDialogScope extends angular.IScope {
+	$close(): void;
+	$dismiss(): void;
+}
+
+export interface IDialogInstance {
+	close(): void;
+	dismiss(): void;
+}
+
+export interface IPromptSettings {
+	acceptHandler(): void;
+	cancelHandler?: {(): void};
+	message: string;
+	okButton?: string;
+	cancelButton?: string;
+}
+
+export interface IAutosaveDialogSettings {
+	scope?: ng.IScope;
+	template?: string;
+	templateUrl?: string;
+	size?: string;
+	data?: any;
+	resolve?: any;
+
+	save: { (...data: any[]): angular.IPromise<void> };
+	triggers?: string;
+}
+
+interface IDialogSettings {
+	scope?: ng.IScope;
+	template?: string;
+	templateUrl?: string;
+	size?: string;
+	data?: any;
+
+	controller?: string;
+	controllerAs?: string;
+	bindToController?: boolean;
+}
+
+export interface IPromptInstance extends IDialogInstance {
+	accept(): void;
+	cancel(): void;
+}
+
+export interface IDialogImplementation<TDialogSettings> {
+	open(options: TDialogSettings, closeHandler?: IDialogCloseHandler): IDialogInstance;
+	prompt(options: IPromptSettings, template: string): IPromptInstance;
+}
