+++ source/services/services.module.ts
@@ -4,7 +4,6 @@ import * as angular from 'angular';
 
 import * as autosave from './autosave/autosave.service';
 import * as autosaveAction from './autosaveAction/autosaveAction.service';
-import * as autosaveDialog from './autosaveDialog/autosaveDialog.module';
 import * as breakpoints from './breakpoints/breakpoints.module';
 import * as componentValidator from './componentValidator/componentValidator.service';
 import * as contentProvider from './contentProvider/contentProvider.service';
@@ -18,7 +17,6 @@ import * as windowWrapper from './windowWrapper/windowWrapper.service';
 export {
 	autosave,
 	autosaveAction,
-	autosaveDialog,
 	breakpoints,
 	componentValidator,
 	contentProvider,
@@ -35,7 +33,6 @@ export var moduleName: string = 'rl.ui.services';
 angular.module(moduleName, [
 	autosave.moduleName,
 	autosaveAction.moduleName,
-	autosaveDialog.moduleName,
 	breakpoints.moduleName,
 	componentValidator.moduleName,
 	contentProvider.moduleName,
